+++ angular-workspace/projects/ni/nimble-angular/src/directives/combobox/tests/nimble-combobox.directive.spec.ts
@@ -69,7 +69,8 @@ describe('Nimble combobox', () => {
                 <nimble-combobox #combobox
                     disabled
                     autocomplete="inline"
-                    error-text="error text">
+                    error-text="error text"
+                    placeholder="Enter value:">
                 </nimble-combobox>`
         })
         class TestHostComponent {
@@ -102,6 +103,11 @@ describe('Nimble combobox', () => {
             expect(nativeElement.autocomplete).toEqual(ComboboxAutocomplete.inline);
         });
 
+        it('will use template string values for placeholder', () => {
+            expect(directive.placeholder).toBe('Enter value:');
+            expect(nativeElement.placeholder).toBe('Enter value:');
+        });
+
         it('will use template string values for errorText', () => {
             expect(directive.errorText).toBe('error text');
             expect(nativeElement.errorText).toBe('error text');
@@ -114,7 +120,8 @@ describe('Nimble combobox', () => {
                 <nimble-combobox #combobox
                     [disabled]="disabled"
                     [autocomplete]="autocomplete"
-                    [error-text]="errorText">
+                    [error-text]="errorText"
+                    [placeholder]="placeholder">
                 </nimble-combobox>
             `
         })
@@ -124,6 +131,7 @@ describe('Nimble combobox', () => {
             public disabled = false;
             public autocomplete: ComboboxAutocomplete = ComboboxAutocomplete.list;
             public errorText = 'initial value';
+            public placeholder = 'Enter value:';
         }
 
         let fixture: ComponentFixture<TestHostComponent>;
@@ -163,6 +171,18 @@ describe('Nimble combobox', () => {
             expect(nativeElement.autocomplete).toEqual(ComboboxAutocomplete.both);
         });
 
+        it('can be configured with property binding for placeholder', () => {
+            expect(directive.placeholder).toEqual('Enter value:');
+            expect(nativeElement.placeholder).toEqual('Enter value:');
+
+            const newPlaceholderValue = 'Enter new value:';
+            fixture.componentInstance.placeholder = newPlaceholderValue;
+            fixture.detectChanges();
+
+            expect(directive.placeholder).toEqual(newPlaceholderValue);
+            expect(nativeElement.placeholder).toEqual(newPlaceholderValue);
+        });
+
         it('can be configured with property binding for errorText', () => {
             expect(directive.errorText).toBe('initial value');
             expect(nativeElement.errorText).toBe('initial value');
@@ -181,6 +201,7 @@ describe('Nimble combobox', () => {
                 <nimble-combobox #combobox
                     [attr.disabled]="disabled"
                     [attr.autocomplete]="autocomplete"
+                    [attr.placeholder]="placeholder"
                     [attr.error-text]="errorText">
                 </nimble-combobox>
             `
@@ -190,6 +211,7 @@ describe('Nimble combobox', () => {
             @ViewChild('combobox', { read: ElementRef }) public elementRef: ElementRef<Combobox>;
             public disabled: BooleanValueOrAttribute = null;
             public autocomplete: ComboboxAutocomplete | undefined = undefined;
+            public placeholder = 'Enter value:';
             public errorText = 'initial value';
         }
 
@@ -228,6 +250,18 @@ describe('Nimble combobox', () => {
 
             expect(directive.autocomplete).toEqual(ComboboxAutocomplete.both);
             expect(nativeElement.autocomplete).toEqual(ComboboxAutocomplete.both);
+        });
+
+        it('can be configured with attribute binding for placeholder', () => {
+            expect(directive.placeholder).toBe('Enter value:');
+            expect(nativeElement.placeholder).toBe('Enter value:');
+
+            const newPlaceholderValue = 'Enter new value:';
+            fixture.componentInstance.placeholder = newPlaceholderValue;
+            fixture.detectChanges();
+
+            expect(directive.placeholder).toEqual(newPlaceholderValue);
+            expect(nativeElement.placeholder).toEqual(newPlaceholderValue);
         });
 
         it('can be configured with attribute binding for errorText', () => {