+++ packages/nimble-components/src/switch/styles.ts
@@ -19,6 +19,7 @@ import {
     fillHoverColor,
     smallDelay
 } from '../theme-provider/design-tokens';
+import { Theme } from '../theme-provider/types';
 import { hexToRgbaCssColor } from '../utilities/style/colors';
 import { focusVisible } from '../utilities/style/focus';
 import { themeBehavior } from '../utilities/style/theme';
@@ -163,8 +164,8 @@ export const styles = css`
     }
 `.withBehaviors(
         themeBehavior(
+            Theme.light,
             css`
-            ${'' /* Light theme */}
             :host {
                 --ni-private-switch-background-disabled-color: ${hexToRgbaCssColor(
                 Black91,
@@ -181,9 +182,11 @@ export const styles = css`
                 0.3
             )};
             }
-        `,
+        `
+        ),
+        themeBehavior(
+            Theme.dark,
             css`
-            ${'' /* Dark theme */}
             :host {
                 --ni-private-switch-background-disabled-color: ${hexToRgbaCssColor(
                 Black15,
@@ -203,9 +206,11 @@ export const styles = css`
                 0.3
             )};
             }
-        `,
+        `
+        ),
+        themeBehavior(
+            Theme.color,
             css`
-            ${'' /* Color theme */}
             :host {
                 --ni-private-switch-background-disabled-color: ${hexToRgbaCssColor(
                 White,